a bit of code emission package refactor
package api refactor
site model classes comments
code fixes after api change
